% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-function.R
\name{clustcureCaseC}
\alias{clustcureCaseC}
\title{Fitting promotion time cure model for clustered failure time data under Case-cohort study.}
\usage{
clustcureCaseC(K, n, Data.1, beta, gam, a, tau, c0, P.BerC, P.BerS, N, boot_n)
}
\arguments{
\item{K}{numeric, the number of cluster.}

\item{n}{numeric, the cluster size.}

\item{Data.1}{a data.frame, the column names of the data.frame are Z.1(Covariate), Z.2(Covariate), T.C(Observation time), delta(Censor indicator),
Delta(Cure indicator, 1 if subject is not cure and 0 otherwise), HH.ij(Whether the covariates of the subject are observable).}

\item{beta}{numeric vector of 3 dimension, initial value of parameters.}

\item{gam}{numeric, the parameter of the Gumbel model.}

\item{a}{numeric, the parameter of covariate Z.2.}

\item{tau}{numeric, the cure threshold.}

\item{c0}{numeric, the parameter of the distribution function for censoring time.}

\item{P.BerC}{numeric, Sampling clusters using Bernoulli sample with probability P.BerC.}

\item{P.BerS}{numeric, Sampling subjects from the selected clusters using Bernoulli sample with probability P.BerS.}

\item{N}{numeric, the degree of the Bernstein basis polynomials.}

\item{boot_n}{numeric, the replication of the weighted bootstrap procedure.}
}
\value{
clustcureCaseC returns a list containing the following components:

\item{coefficients}{the estimators of coefficients.}

\item{se}{the standard errors of the coefficients estimators.}
}
\description{
This function fits a promotion time cure model for clustered failure time data under Case-cohort study, where the unknown baseline distribution function is
approximated by Bernstein polynomial. In this case-cohort design, the covariates can be considered as missing at random. This function employs the inverse probability
weighting to construct likelihood function.
}
\examples{
library(maxLik)
## true parameter: c(1,1,-2)
K<-400; n<-6
data(sampleData)
beta<-c(1,1,-2); gam<-0.2; a<-0.15; tau<-1; c0<-4
P.BerC<-0.5; P.BerS<-0.5; N<-4; boot_n<-3
## estimate parameters
# clustcureCaseC(K,n,sampleData,beta,gam,a,tau,c0,P.BerC,P.BerS,N,boot_n)
}
