% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-function.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{sample data}
\format{
A data frame with ... rows and ... variables, including:
\describe{
\item{id}{the clustering number}
\item{Z.1}{covariate 1 comes from b(1,0.5) distribution}
\item{Z.2}{covariate 2 comes from U(a,a+1) distribution}
\item{T.c}{the observed failure time}
\item{delta}{the censoring indicator}}
\item{Delta}{cured indicator}
\item{HH.ij}{the case-cohort indicator, 1 indicating the subject from the case-cohort sample}
}
}
\usage{
sampleData
}
\description{
this is a sample data
}
\keyword{datasets}
